feat: rewrite documentation
